feat: refactor pagination and categories components for improved structure and functionality
